Add TrainConfig struct and include it in TrainInstance and TrainManager
